requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
docxtpl==0.16.7
python-docx==0.8.11
reportlab==4.0.4
jinja2==3.1.2
weasyprint==58.1 